--- mcomp_Java/src/member/ui/package-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * @author David Avery
- *
- */
-package member.ui;
